+++ core/application/impl/configuration_storage_impl.cpp
@@ -30,7 +30,7 @@ namespace kagome::application {
     return config_.authorities;
   }
 
-  uint32_t ConfigurationStorageImpl::getExtrinsicApiPort() const {
+  uint16_t ConfigurationStorageImpl::getExtrinsicApiPort() const {
     return config_.api_ports.extrinsic_api_port;
   }
