+++ core/application/impl/configuration_storage_impl.hpp
@@ -23,7 +23,7 @@ namespace kagome::application {
     std::vector<crypto::SR25519PublicKey> getSessionKeys() const override;
     std::vector<crypto::ED25519PublicKey> getAuthorities() const override;
 
-    uint32_t getExtrinsicApiPort() const override;
+    uint16_t getExtrinsicApiPort() const override;
 
    private:
     KagomeConfig config_;
